# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Forum Group Discussion",
        "title": "Forum Group Discussion API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Kelompok 22",
            "url": "http://www.swagger.io/support",
            "email": "erikriosetiawan15@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "moot-rest-api.herokuapp.com",
    "basePath": "/api/v1",
    "paths": {
...
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "This endpoint is used to get all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get Users",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page, default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit, default 20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "options: registered_date, ranking, default registered_date",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "options: active, banned, default active",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.profilesResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint is used to get their own user profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get Own Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.profileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{username}": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "This endpoint is used to get the another user by username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User by Username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.profileResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{username}/banned": {
            "put": {
                "security": [
                    {
                        "ApiKey": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint is used to banned/unbanned a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Banned/Unbanned a User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{username}/follow": {
            "put": {
                "security": [
                    {
                        "ApiKey": []
                    },
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "This endpoint is used to follow/unfollow a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Follow/Unfollow a User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{username}/threads": {
            "get": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "description": "This endpoint is used to get the user threads",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get User Threads",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page, default 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit, default 10",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.threadsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/echo.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "controller.pageInfoData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer",
                    "x-order": "0"
                },
                "page": {
                    "type": "integer",
                    "x-order": "1"
                },
                "pageTotal": {
                    "type": "integer",
                    "x-order": "2"
                },
                "total": {
                    "type": "integer",
                    "x-order": "3"
                }
            }
        },
        "controller.profileData": {
            "type": "object",
            "properties": {
                "userID": {
                    "type": "string",
                    "x-order": "0"
                },
                "username": {
                    "type": "string",
                    "x-order": "1"
                },
                "email": {
                    "type": "string",
                    "x-order": "2"
                },
                "name": {
                    "type": "string",
                    "x-order": "3"
                },
                "role": {
                    "type": "string",
                    "x-order": "4"
                },
                "isActive": {
                    "type": "boolean",
                    "x-order": "5"
                },
                "registeredOn": {
                    "description": "RegisteredOn layout format: time.RFC822 (02 Jan 06 15:04 MST)",
                    "type": "string",
                    "x-order": "6"
                },
                "totalThread": {
                    "type": "integer",
                    "x-order": "7"
                },
                "totalFollower": {
                    "type": "integer",
                    "x-order": "8"
                },
                "isFollowed": {
                    "type": "boolean",
                    "x-order": "9"
                }
            }
        },
        "controller.profileResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "x-order": "0"
                },
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "x-order": "2",
                    "$ref": "#/definitions/controller.profileData"
                }
            }
        },
        "controller.profilesInfoWrapper": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.profileData"
                    },
                    "x-order": "0"
                },
                "pageInfo": {
                    "x-order": "1",
                    "$ref": "#/definitions/controller.pageInfoData"
                }
            }
        },
        "controller.profilesResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "x-order": "0"
                },
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "x-order": "2",
                    "$ref": "#/definitions/controller.profilesInfoWrapper"
                }
            }
        },
...
        "controller.threadsData": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string",
                    "x-order": "0"
                },
                "title": {
                    "type": "string",
                    "x-order": "1"
                },
                "categoryID": {
                    "type": "string",
                    "x-order": "10"
                },
                "categoryName": {
                    "type": "string",
                    "x-order": "11"
                },
                "publishedOn": {
                    "description": "PublishedOn layout format: time.RFC822 (02 Jan 06 15:04 MST)",
                    "type": "string",
                    "x-order": "12"
                },
                "isLiked": {
                    "type": "boolean",
                    "x-order": "13"
                },
                "isFollowed": {
                    "type": "boolean",
                    "x-order": "14"
                },
                "description": {
                    "type": "string",
                    "x-order": "2"
                },
                "totalViewer": {
                    "type": "integer",
                    "x-order": "3"
                },
                "totalLike": {
                    "type": "integer",
                    "x-order": "4"
                },
                "totalFollower": {
                    "type": "integer",
                    "x-order": "5"
                },
                "totalComment": {
                    "type": "integer",
                    "x-order": "6"
                },
                "creatorID": {
                    "type": "string",
                    "x-order": "7"
                },
                "creatorUsername": {
                    "type": "string",
                    "x-order": "8"
                },
                "creatorName": {
                    "type": "string",
                    "x-order": "9"
                }
            }
        },
        "controller.threadsInfoWrapper": {
            "type": "object",
            "properties": {
                "threads": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.threadsData"
                    },
                    "x-order": "0"
                },
                "pageInfo": {
                    "x-order": "1",
                    "$ref": "#/definitions/controller.pageInfoData"
                }
            }
        },
        "controller.threadsResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "x-order": "0"
                },
                "message": {
                    "type": "string",
                    "x-order": "1"
                },
                "data": {
                    "x-order": "2",
                    "$ref": "#/definitions/controller.threadsInfoWrapper"
                }
            }
        },
...
        "echo.HTTPError": {
            "type": "object",
            "properties": {
                "message": {}
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKey": {
            "type": "apiKey",
            "name": "API-Key",
            "in": "header"
        },
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}